条件分岐の修正

+++ app/controllers/items_controller.rb
@@ -36,9 +36,12 @@ class ItemsController < ApplicationController
   end  
 
   def destroy
-    item = Item.find(params[:id])
-    item.destroy
-    redirect_to root_path
+    if item = Item.find(params[:id])
+      item.destroy
+      redirect_to root_path
+    else
+      redirect_to root_path
+    end    
   end  
   private
